--- components/admin/layout/AnimatedPageWrapper.tsx
@@ -1,383 +1,0 @@
-'use client';
-
-import { motion, AnimatePresence, type Variants } from 'framer-motion';
-import { usePathname } from 'next/navigation';
-
-/* ============================================
-   PAGE TRANSITION VARIANTS
-   ============================================ */
-
-const pageVariants: Variants = {
-  initial: {
-    opacity: 0,
-    y: 20,
-    filter: 'blur(10px)',
-    scale: 0.98,
-  },
-  animate: {
-    opacity: 1,
-    y: 0,
-    filter: 'blur(0px)',
-    scale: 1,
-    transition: {
-      duration: 0.4,
-      ease: [0.22, 1, 0.36, 1] as [number, number, number, number],
-      staggerChildren: 0.1,
-    },
-  },
-  exit: {
-    opacity: 0,
-    y: -20,
-    filter: 'blur(10px)',
-    scale: 0.98,
-    transition: {
-      duration: 0.3,
-      ease: [0.4, 0, 1, 1] as [number, number, number, number],
-    },
-  },
-};
-
-const contentVariants: Variants = {
-  initial: {
-    opacity: 0,
-    y: 10,
-  },
-  animate: {
-    opacity: 1,
-    y: 0,
-    transition: {
-      duration: 0.3,
-      ease: [0, 0, 0.2, 1] as [number, number, number, number],
-    },
-  },
-};
-
-/* ============================================
-   ANIMATED PAGE WRAPPER
-   ============================================ */
-
-interface AnimatedPageWrapperProps {
-  children: React.ReactNode;
-  className?: string;
-  delay?: number;
-}
-
-export function AnimatedPageWrapper({
-  children,
-  className = '',
-  delay: _delay = 0,
-}: AnimatedPageWrapperProps) {
-  const pathname = usePathname();
-
-  return (
-    <AnimatePresence mode="wait">
-      <motion.div
-        key={pathname}
-        variants={pageVariants}
-        initial="initial"
-        animate="animate"
-        exit="exit"
-        className={className}
-        style={{
-          transformOrigin: 'center top',
-        }}
-      >
-        {children}
-      </motion.div>
-    </AnimatePresence>
-  );
-}
-
-/* ============================================
-   ANIMATED CONTENT SECTION
-   ============================================ */
-
-interface AnimatedContentProps {
-  children: React.ReactNode;
-  className?: string;
-  delay?: number;
-}
-
-export function AnimatedContent({
-  children,
-  className = '',
-  delay = 0,
-}: AnimatedContentProps) {
-  return (
-    <motion.div
-      variants={contentVariants}
-      initial="initial"
-      animate="animate"
-      transition={{ delay }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   STAGGER CONTAINER
-   ============================================ */
-
-interface StaggerContainerProps {
-  children: React.ReactNode;
-  className?: string;
-  staggerDelay?: number;
-  initialDelay?: number;
-}
-
-export function StaggerContainer({
-  children,
-  className = '',
-  staggerDelay = 0.1,
-  initialDelay = 0.2,
-}: StaggerContainerProps) {
-  return (
-    <motion.div
-      initial="initial"
-      animate="animate"
-      variants={{
-        initial: { opacity: 0 },
-        animate: {
-          opacity: 1,
-          transition: {
-            staggerChildren: staggerDelay,
-            delayChildren: initialDelay,
-          },
-        },
-      }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   STAGGER ITEM
-   ============================================ */
-
-interface StaggerItemProps {
-  children: React.ReactNode;
-  className?: string;
-  direction?: 'up' | 'down' | 'left' | 'right';
-}
-
-export function StaggerItem({
-  children,
-  className = '',
-  direction = 'up',
-}: StaggerItemProps) {
-  const directionMap = {
-    up: { y: 20, x: 0 },
-    down: { y: -20, x: 0 },
-    left: { x: 20, y: 0 },
-    right: { x: -20, y: 0 },
-  };
-
-  const offset = directionMap[direction];
-
-  const itemVariants: Variants = {
-    initial: {
-      opacity: 0,
-      ...offset,
-    },
-    animate: {
-      opacity: 1,
-      x: 0,
-      y: 0,
-      transition: {
-        duration: 0.4,
-        ease: [0.25, 0.4, 0.25, 1] as [number, number, number, number],
-      },
-    },
-  };
-
-  return (
-    <motion.div
-      variants={itemVariants}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   FADE IN WRAPPER
-   ============================================ */
-
-interface FadeInProps {
-  children: React.ReactNode;
-  className?: string;
-  delay?: number;
-  duration?: number;
-}
-
-export function FadeIn({
-  children,
-  className = '',
-  delay = 0,
-  duration = 0.4,
-}: FadeInProps) {
-  return (
-    <motion.div
-      initial={{ opacity: 0 }}
-      animate={{ opacity: 1 }}
-      transition={{
-        delay,
-        duration,
-        ease: 'easeOut',
-      }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   SLIDE IN WRAPPER
-   ============================================ */
-
-interface SlideInProps {
-  children: React.ReactNode;
-  className?: string;
-  direction?: 'up' | 'down' | 'left' | 'right';
-  delay?: number;
-  duration?: number;
-  distance?: number;
-}
-
-export function SlideIn({
-  children,
-  className = '',
-  direction = 'up',
-  delay = 0,
-  duration = 0.4,
-  distance = 30,
-}: SlideInProps) {
-  const directionMap = {
-    up: { y: distance, x: 0 },
-    down: { y: -distance, x: 0 },
-    left: { x: distance, y: 0 },
-    right: { x: -distance, y: 0 },
-  };
-
-  const offset = directionMap[direction];
-
-  return (
-    <motion.div
-      initial={{ opacity: 0, ...offset }}
-      animate={{ opacity: 1, x: 0, y: 0 }}
-      transition={{
-        delay,
-        duration,
-        ease: [0.25, 0.4, 0.25, 1],
-      }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   SCALE IN WRAPPER
-   ============================================ */
-
-interface ScaleInProps {
-  children: React.ReactNode;
-  className?: string;
-  delay?: number;
-  duration?: number;
-  initialScale?: number;
-}
-
-export function ScaleIn({
-  children,
-  className = '',
-  delay = 0,
-  duration = 0.4,
-  initialScale = 0.95,
-}: ScaleInProps) {
-  return (
-    <motion.div
-      initial={{ opacity: 0, scale: initialScale }}
-      animate={{ opacity: 1, scale: 1 }}
-      transition={{
-        delay,
-        duration,
-        ease: [0.25, 0.4, 0.25, 1],
-      }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   BLUR IN WRAPPER
-   ============================================ */
-
-interface BlurInProps {
-  children: React.ReactNode;
-  className?: string;
-  delay?: number;
-  duration?: number;
-}
-
-export function BlurIn({
-  children,
-  className = '',
-  delay = 0,
-  duration = 0.5,
-}: BlurInProps) {
-  return (
-    <motion.div
-      initial={{ opacity: 0, filter: 'blur(10px)' }}
-      animate={{ opacity: 1, filter: 'blur(0px)' }}
-      transition={{
-        delay,
-        duration,
-        ease: [0.25, 0.4, 0.25, 1],
-      }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   PAGE HEADER ANIMATION
-   ============================================ */
-
-interface PageHeaderProps {
-  children: React.ReactNode;
-  className?: string;
-}
-
-export function AnimatedPageHeader({ children, className = '' }: PageHeaderProps) {
-  return (
-    <motion.div
-      initial={{ opacity: 0, y: -20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{
-        duration: 0.5,
-        ease: [0.22, 1, 0.36, 1],
-      }}
-      className={className}
-    >
-      {children}
-    </motion.div>
-  );
-}
-
-/* ============================================
-   EXPORTS
-   ============================================ */
-
-export default AnimatedPageWrapper;
