+++ components/admin/layout/index.ts
@@ -1,21 +1,5 @@
 // Admin Layout Components
 
-// Background Effects
-export { default as AdminBackgroundEffects, ParticleField, MatrixRain, FloatingOrb, GridPattern, GradientOverlay } from './AdminBackgroundEffects';
-
-// Page Animations
-export {
-  AnimatedPageWrapper,
-  AnimatedContent,
-  StaggerContainer,
-  StaggerItem,
-  FadeIn,
-  SlideIn,
-  ScaleIn,
-  BlurIn,
-  AnimatedPageHeader
-} from './AnimatedPageWrapper';
-
 // Toast System
 export {
   ToastContainer,
@@ -24,14 +8,14 @@ export {
   toast,
   setGlobalToastHandler,
   type Toast,
-  type ToastType
-} from './ToastNeon';
+  type ToastType,
+} from "./ToastNeon";
 
 // Sidebar
-export { AdminSidebarImmersive } from './AdminSidebarImmersive';
+export { AdminSidebarImmersive } from "./AdminSidebarImmersive";
 
 // Header
-export { AdminHeaderImmersive } from './AdminHeaderImmersive';
+export { AdminHeaderImmersive } from "./AdminHeaderImmersive";
 
 // Command Palette
-export { AdminCommandPalette, useCommandPalette } from './AdminCommandPalette';
+export { AdminCommandPalette, useCommandPalette } from "./AdminCommandPalette";
