--- layout/wrapper.jsx
@@ -1,5 +1,0 @@
-const Wrapper = ({ children }) => {
-  return <>{children}</>;
-};
-
-export default Wrapper;
