+++ package.json
@@ -51,7 +51,6 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@emailjs/browser": "^4.4.1",
     "@hookform/resolvers": "^5.2.2",
     "@prisma/client": "^6.19.0",
     "@radix-ui/react-avatar": "^1.1.11",
@@ -69,8 +68,6 @@
     "@radix-ui/react-toast": "^1.2.15",
     "@radix-ui/react-tooltip": "^1.2.8",
     "@react-pdf/renderer": "^4.3.1",
-    "@react-three/drei": "^10.7.7",
-    "@react-three/fiber": "^9.4.0",
     "@tiptap/extension-link": "^3.11.0",
     "@tiptap/extension-placeholder": "^3.11.0",
     "@tiptap/react": "^3.11.0",
@@ -86,10 +83,7 @@
     "eslint": "9",
     "eslint-config-next": "^16.0.3",
     "framer-motion": "^12.23.24",
-    "isomorphic-dompurify": "^2.33.0",
     "lucide-react": "^0.554.0",
-    "maath": "^0.10.8",
-    "mapbox-gl": "^3.16.0",
     "next": "^16.0.7",
     "next-intl": "^4.5.5",
     "next-themes": "^0.2.1",
@@ -97,10 +91,8 @@
     "react-dom": "^19.2.0",
     "react-dropzone": "^14.3.8",
     "react-hook-form": "^7.66.1",
-    "react-map-gl": "^8.1.0",
     "react-toastify": "^11.0.5",
     "tailwind-merge": "^3.4.0",
-    "three": "^0.181.2",
     "xlsx": "^0.18.5",
     "zod": "^4.1.12"
   },
@@ -111,7 +103,6 @@
     "@types/node": "24.10.1",
     "@types/react": "19.2.6",
     "@types/react-dom": "19.2.3",
-    "@types/three": "^0.181.0",
     "autoprefixer": "^10.4.22",
     "dotenv-cli": "^11.0.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
