--- package.json.backup
@@ -1,42 +1,0 @@
-{
-  "name": "tokyo-nextjs",
-  "version": "0.1.0",
-  "private": true,
-  "engines": {
-    "node": "22"
-  },
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@emailjs/browser": "^3.10.0",
-    "aos": "^2.3.4",
-    "eslint": "8.31.0",
-    "eslint-config-next": "13.1.1",
-    "next": "^15.2.5",
-    "next-share": "^0.19.0",
-    "next-themes": "^0.2.1",
-    "photoswipe": "^5.3.4",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-icons": "^4.7.1",
-    "react-map-gl": "^5.2.11",
-    "react-modal": "^3.16.1",
-    "react-modal-video": "^2.0.0",
-    "react-photoswipe-gallery": "^2.2.2",
-    "react-responsive": "^9.0.2",
-    "react-slick": "^0.29.0",
-    "react-tabs": "^6.0.0",
-    "react-toastify": "^9.1.1",
-    "react-tooltip": "^4.2.21",
-    "react-typed": "^1.2.0",
-    "sass": "^1.57.1",
-    "slick-carousel": "^1.8.1"
-  },
-  "devDependencies": {
-    "sass-migrator": "^2.3.1"
-  }
-}
