--- utilis/linkActiveChecker.js
@@ -1,6 +1,0 @@
-// is active link check
-export const isActiveLink = (menuPath, routePath) => {
-  if (menuPath && routePath) {
-    return menuPath.replace(/\/\d+/, "") === routePath.replace(/\/\d+/, "");
-  }
-};
